clients/mqtt-nats/requirements.txt
paho-mqtt>=2.1.0
nats-py>=2.7.2
